BUGFIX: Show build message when file not in view
If there was a build error in a file not being edited, the error
panel would not be shown (and a strack trace would be dumped on the
console). This was due to an attempt to set the error markers on
an non-existent view.

It was the underlying error in #8