Use continuations for request/response handling
Leverages on a new feature of stack-ide (protocol 0.1.1) where
we can tag requests with a sequence code and get it back on the response
to add, to each request, the action we want to perform, asynchronously,
when the response comes back.

This opens the way to easily implementing features such as: show the
type of expression on a hot-key, etc.